feature #112: update CuboidDTO methods to return long values for dimensions and areas

+++ core/src/main/java/cn/lunadeer/dominion/misc/Asserts.java
@@ -406,7 +406,7 @@ public class Asserts {
         if (bypassLimit(associatedPlayer)) {
             return;
         }
-        int amount;
+        long amount;
         if (ecoConf.squareOnly) {
             amount = after.minusSquareWith(before);
         } else {
